Add maps API key
Note: It's gross to have this sotred in the repo,
but it's the only way to get the key into the api
with github pages. The key is locked down by domain
to mitigate it being in the codebase. This does mean
than local development environments google map instances
will through an error.